Fix aria-live bug
The transition previously applied to all CSS properties of
the element. This prevented aria-live from announcing changes
to the notification section because the JS to add the new content
finished executing before the transition completed. By the time
the new content was added, the notification section was still not
yet a part of the accessibility tree.
This change limits the transition to only the property it was
intended for (i.e. to slide the element on-screen from the left).